kill web content process after each Navigation test (#1801)
Task/Issue URL: https://app.asana.com/1/137249556945/project/1205237866452338/task/1211041653522693?focus=true

### Description
- Terminate web content process after each Navigation test run to reduce
the resource consumption

+++ SharedPackages/BrowserServicesKit/Tests/NavigationTests/Helpers/WKWebView+KillWebContentProcess.swift
@@ -22,9 +22,9 @@ extension WKWebView {
 
     func killWebContentProcess() {
         let webContentProcessInfo = (WKProcessPool.perform(Selector(("_webContentProcessInfo"))).takeUnretainedValue() as? [NSObject])!
-        let processInfo = webContentProcessInfo.first {
-            ($0.value(forKey: "webViews") as? [WKWebView])!.contains(self)
-        }!
+        guard let processInfo = webContentProcessInfo.first(where: {
+            ($0.value(forKey: "webViews") as? [WKWebView])?.contains(self) == true
+        }) else { return }
         let pid = processInfo.value(forKey: "pid") as! pid_t
 
         kill(pid, SIGTERM)
